优化过滤器，实现gateway filter按需配置
